fix(admin): gate /admin in the layout, drop broken middleware self-fetch
The /admin panel was 500ing because the middleware was doing an
HTTPS fetch back to itself (request.nextUrl.origin +
/api/auth/get-session) and that fetch was failing with
ERR_SSL_PACKET_LENGTH_TOO_LONG. The error happened on every admin
request — middleware threw before ever reaching the page.

Move the gate to a server-component layout that calls
auth.api.getSession() directly via the existing getSession() helper.
No HTTP roundtrip, no TLS, no self-fetch. The previous client UI
moves to AdminLayoutClient unchanged.

The middleware was only handling /admin, so the file goes away.
Note: api/admin/* routes were never gated by this middleware
(matcher was /admin/:path*, not /api/admin/:path*). They still
aren't — that's a separate hardening pass.
