--- middleware.ts
@@ -1,47 +1,0 @@
-import { betterFetch } from "@better-fetch/fetch";
-import { NextResponse } from "next/server";
-import type { NextRequest } from "next/server";
-
-interface Session {
-  session: {
-    id: string;
-    userId: string;
-    expiresAt: Date;
-  };
-  user: {
-    id: string;
-    email: string;
-    isAdmin?: boolean;
-  };
-}
-
-export async function middleware(request: NextRequest) {
-  // Check if the route is an admin route
-  if (request.nextUrl.pathname.startsWith("/admin")) {
-    const { data: session } = await betterFetch<Session>(
-      "/api/auth/get-session",
-      {
-        baseURL: request.nextUrl.origin,
-        headers: {
-          cookie: request.headers.get("cookie") || "",
-        },
-      }
-    );
-
-    if (!session) {
-      return NextResponse.redirect(new URL("/auth/login", request.url));
-    }
-
-    // Check if user is admin
-    if (!session.user.isAdmin) {
-      return NextResponse.redirect(new URL("/", request.url));
-    }
-  }
-
-  return NextResponse.next();
-}
-
-// Configure which routes to run middleware on
-export const config = {
-  matcher: ["/admin/:path*"],
-};
